refactor syft handlers module layout
